requirements.txt
# Requisitos principais
mem0>=0.1.65
openai>=1.33.0
python-dotenv>=1.0.0

# Banco de dados e armazenamento vetorial
pgvector>=0.3.0
sqlalchemy>=2.0.0
supabase>=2.0.0
psycopg2-binary>=2.9.5
vecs>=0.2.6

# Utilitários
pydantic>=2.7.0
requests>=2.0.0
tqdm>=4.0.0
numpy>=1.20.0

# Processamento de linguagem natural
nltk>=3.8.1
scikit-learn>=1.3.2

# Logging e CLI
loguru>=0.7.0
colorama>=0.4.6

# Interface web
streamlit>=1.32.0
streamlit-chat>=0.1.1
extra-streamlit-components>=0.1.60

# Testes
pytest>=7.4.0
pytest-cov>=4.1.0
